update structure files for admin
