+++ src/components/layout/AdminLayout.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const AdminLayout = () => {
+  return (
+    <div>AdminLayout</div>
+  )
+}
+
+export default AdminLayout
