--- src/pages/ServicePage.jsx
@@ -1,14 +1,0 @@
-import React from "react";
-
-import { ServicesGrid } from "../components/services";
-import { LayoutPage } from "../components/layout";
-
-const BookingRoom = () => {
-  return (
-    <LayoutPage>
-      <ServicesGrid />
-    </LayoutPage>
-  );
-};
-
-export default BookingRoom;
