+++ src/pages/customer/ServicePage.jsx
@@ -1,0 +1,14 @@
+import React from "react";
+
+import { ServicesGrid } from "../../components/services";
+import { CustomerLayout } from "../../components/layout";
+
+const BookingRoom = () => {
+  return (
+    <CustomerLayout>
+      <ServicesGrid />
+    </CustomerLayout>
+  );
+};
+
+export default BookingRoom;
