+++ src/pages/customer/index.js
@@ -1,0 +1,4 @@
+export { default as Home } from './Home'
+export { default as BookingRoom } from './BookingRoom'
+export { default as ServicePage } from './ServicePage'
+export { default as RoomDetail } from './RoomDetail'
