--- src/pages/index.js
@@ -1,3 +1,0 @@
-export { default as Home } from './Home'
-export { default as BookingRoom } from './BookingRoom'
-export { default as ServicePage } from './ServicePage'
